pyproject.toml
[project]
name = "spy-cat-agency-api"
version = "0.1.0"
description = "FastAPI CRUD assessment for Spy Cat Agency (cats, missions, targets) with TheCatAPI breed validation."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.1",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]
[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [".venv", "__pycache__", "sca.db"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]


[tool.ruff.lint.per-file-ignores]
"app/db/models.py" = ["UP006", "UP035"]
"app/api/*.py" = ["B008"]
